+++ src/main/java/gragongit/arcaneartistry/common/network/StaffRenderOffsetPayload.java
@@ -6,12 +6,12 @@ import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 
-public record StaffRenderOffsetPayload(double yaw, double pitch) implements CustomPacketPayload {
+public record StaffRenderOffsetPayload(float yaw, float pitch) implements CustomPacketPayload {
   public static final CustomPacketPayload.Type<StaffRenderOffsetPayload> TYPE =
       new CustomPacketPayload.Type<>(ArcaneArtistry.id("staff_render_offset"));
 
   public static final StreamCodec<FriendlyByteBuf, StaffRenderOffsetPayload> CODEC = StreamCodec
-      .composite(ByteBufCodecs.DOUBLE, StaffRenderOffsetPayload::yaw, ByteBufCodecs.DOUBLE, StaffRenderOffsetPayload::pitch,
+      .composite(ByteBufCodecs.FLOAT, StaffRenderOffsetPayload::yaw, ByteBufCodecs.FLOAT, StaffRenderOffsetPayload::pitch,
           StaffRenderOffsetPayload::new);
 
   @Override
